refactor the structure of the Graph component
